+++ CHANGELOG.md
@@ -8,6 +8,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.27.1] - 2025-06-01
 MR #28
 
+### Added
+- pre-commit and sreamlit_searchbox to dependencies
+
+### Changed
+- Transport colors are the same used by RATP/SNCF
+
 ### Fixed
 - Weighting processing for congestion awareness in the graph model
 - Display for transfers